requirements.txt
Django>=4.2,<5.0
python-dotenv>=1.0.0
django-allauth>=0.60.0
stripe>=7.0.0
psycopg2-binary>=2.9.9
dj-database-url>=2.3.0
gunicorn>=21.0.0
whitenoise>=6.0.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
python-decouple>=3.8
PyJWT>=2.10.0
cryptography>=44.0.0
Pillow>=10.0.0 
djangorestframework>=3.14.0 